requirements.txt
pandas
openai
requests
python-dotenv
backoff
numpy
matplotlib
scikit-learn
